+++ example/assets.js
@@ -1,0 +1,7 @@
+import { instantiate, invoke } from "./assets.mjs";
+import dartModule from "./assets.wasm";
+
+const dartInstance = await instantiate(dartModule);
+invoke(dartInstance);
+
+export default __fetcher;