+++ lib/http.dart
@@ -1,4 +1,0 @@
-library;
-
-export 'src/http/types.dart';
-export 'src/http/extern.dart';