+++ lib/workers.dart
@@ -1,3 +1,0 @@
-library;
-
-export 'src/workers.dart';