+++ web/exec.d.ts
@@ -1,1 +1,0 @@
-export function exec(fn: () => void, req: Request): Promise<Response>;