+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 2025.5.15 Braking Change
+
+```dart
+// example/simple.dart
+import 'package:poor_web/js.dart';
+import 'package:cf_workers/cf_workers.dart';
+
+main() => fetch(_handler);
+
+Future<Response> _handler(Request request, Env env, Context ctx) async =>
+    Response(String('Hello, Dart!'));
+```
+
+```js
+// example/simple.js
+import { instantiate, invoke } from "./simple.mjs";
+import dartModule from "./simple.wasm";
+
+const dartInstance = await instantiate(dartModule);
+invoke(dartInstance);
+
+export default __dart_fetch;
+```
+
 ## 1.0.0-rc.2
 
 Changed Handler IO.